Move LinkRules into SharedRules
